Move Shift to ast::traits
