+++ lang/syntax/src/ast/mod.rs
@@ -3,7 +3,7 @@ mod exp;
 mod ident;
 mod lookup;
 pub mod lookup_table;
-mod traits;
+pub mod traits;
 
 pub use decls::*;
 pub use exp::*;
