+++ lang/syntax/src/ast/traits/mod.rs
@@ -1,9 +1,11 @@
 mod has_span;
 mod has_type;
 mod occurs;
+mod shift;
 pub mod subst;
 
 pub use has_span::*;
 pub use has_type::*;
 pub use occurs::*;
+pub use shift::*;
 pub use subst::*;
