+++ lang/syntax/src/ctx/values.rs
@@ -4,6 +4,7 @@
 
 use std::rc::Rc;
 
+use crate::ast::traits::Shift;
 use crate::ast::*;
 use crate::common::*;
 use crate::ctx::{Context, LevelCtx};
